move invite view to users app
