+++ commcare_connect/opportunity/urls.py
@@ -8,7 +8,6 @@ from commcare_connect.opportunity.views import (
     OpportunityUserLearnProgress,
     OpportunityUserTableView,
     OpportunityUserVisitTableView,
-    accept_invite,
     add_budget_existing_users,
     download_visit_export,
     export_user_visits,
@@ -33,7 +32,6 @@ urlpatterns = [
         view=OpportunityUserLearnProgress.as_view(),
         name="user_learn_progress",
     ),
-    path("accept_invite/<slug:invite_id>", view=accept_invite, name="accept_invite"),
     path(
         "<int:pk>/add_budget",
         view=add_budget_existing_users,
